Extract out the list into its own component (NotesList.tsx)
